queue-guard/requirements.txt
pika==1.1.0
ruamel.yaml==0.16.10
requests==2.22.0
pyrabbit==1.1.0
pyzmq==18.1.1